requirements.txt
yfinance==0.1.59
google-auth==1.21.2
google-cloud-storage==1.38.0
pandas==1.0.5
gcsfs==2021.6.0
Flask==1.1.2
gunicorn==20.1.0
pytest==5.4.3


